+++ apps/jaspr_pad/lib/components/elements/markdown.imports.dart
@@ -4,4 +4,6 @@
 // GENERATED FILE, DO NOT MODIFY
 // Generated with jaspr_builder
 
-export '../../generated/imports/_web.dart' if (dart.library.io) '../../generated/imports/_stubs.dart' show highlightAll;
+export '../../generated/imports/_web.dart'
+    if (dart.library.io) '../../generated/imports/_stubs.dart'
+    show highlightAll;