+++ packages/jaspr_lints/lib/src/lints/styles_ordering_lint.dart
@@ -110,7 +110,7 @@ class OrderStylesFix extends DartFix {
     Diagnostic analysisError,
     List<Diagnostic> others,
   ) {
-    if (analysisError.data case (NodeList<Expression> arguments, List<String> params)) {
+    if (analysisError.data case (NodeList<Expression> arguments, List<String?> params)) {
       var violatesOrder = !StylesOrderingLint.isOrdered(arguments, params);
       if (!violatesOrder) {
         return;