+++ packages/jaspr/CHANGELOG.md
@@ -1,4 +1,8 @@
-## Unreleased minor
+## 0.21.1
+
+- Fix rendering issue with fragments.
+
+## 0.21.0
 
 **Read the full [Release Notes](https://docs.jaspr.site/releases/v/0.21.0) for this version.**