+++ packages/jaspr_flutter_embed/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.4.9
+
+- `jaspr` upgraded to `0.21.1`
+
+## 0.4.8
+
+- `jaspr` upgraded to `0.21.0`
+
 ## 0.4.7
 
 - `jaspr` upgraded to `0.20.0`