+++ packages/jaspr_router/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.7.3
+
+- `jaspr` upgraded to `0.21.1`
+- `jaspr_test` upgraded to `0.21.1`
+
+## 0.7.2
+
+- `jaspr` upgraded to `0.21.0`
+- `jaspr_test` upgraded to `0.21.0`
+
 ## 0.7.1
 
 - `jaspr` upgraded to `0.20.0`